// src/lib/clients/myanimelist/generated/models/UserAnimeListEdge.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { AnimeForList } from './AnimeForList';
import type { AnimeListStatus } from './AnimeListStatus';

export type UserAnimeListEdge = {
    node?: AnimeForList;
    list_status?: AnimeListStatus;
};